requirements.txt
# Core dependencies
streamlit==1.39.0
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.0

# OpenAI
openai==1.12.0

# Data & APIs
yfinance==0.2.44
alpha-vantage==2.3.1
newsapi-python==0.2.7
requests==2.32.3

# Data analysis
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1

# Visualization
plotly==5.24.1
matplotlib==3.9.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Google Sheets Integration (optional)
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
